refactor: replace `u32` worldwide day with `WorldwideDay` type across crates

+++ crates/core/metadosis/src/settlement.rs
@@ -261,7 +261,7 @@ fn dispatch_brief(
     let brief_supply = if is_green { supply } else { U256::ZERO };
     let receipt = outbe_desis::api::dispatch_auction_brief(
         ctx.storage.clone(),
-        u32::from(wwd),
+        wwd,
         brief_supply,
         entry_price,
         is_green,
